+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added support for quiet alert checks
+- Added support for baselime report slack
 
 ## [0.0.41] 2022-12-23